% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildModel_functions.R
\name{BuildModel}
\alias{BuildModel}
\title{Fit a regression model}
\usage{
BuildModel(output, degree = 1)
}
\arguments{
\item{output}{ConDecon object with low dimensional embedding of distributions of cell abundances and correlations}

\item{degree}{degree of the polynomial regression model (default = 1)}
}
\value{
ConDecon object with fit model
}
\description{
Learn a model that maps between the space of cell abundances and correlations
}
\examples{
data(counts_gps)
data(latent_gps)
data(variable_genes_gps)
data(bulk_gps)

# For this example, we will reduce the training size to max.iter = 50 to reduce run time
TrainingSet = BuildTrainingSet(count = counts_gps, latent = latent_gps, max.iter = 50)
ConDecon_obj = Map2Latent(TrainingSet = TrainingSet, latent = latent_gps, count = counts_gps,
bulk = bulk_gps, variable.features = variable_genes_gps)

ConDecon_obj = BuildModel(ConDecon_obj)
}
