% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_gps}
\alias{bulk_gps}
\title{Example bulk data}
\format{
A matrix with 5,000 rows (genes) and 5 columns (bulk samples)
}
\source{
{Simulated bulk data generated by aggregating cell types/groups (gps) in counts_gps}
}
\usage{
bulk_gps
}
\description{
Example bulk gene expression data generated by aggregating cells from counts_gps single-cell data.
}
\examples{
data(bulk_gps)
}
\keyword{datasets}
