% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTrainingSet_functions.R
\name{BulkCells}
\alias{BulkCells}
\title{Aggregate cells in bulk data based on the Gaussian kernal}
\usage{
BulkCells(pick.cells, max.iter, step, max.cent, tot.cells, n, count, verbose)
}
\arguments{
\item{count}{}
}
\value{

}
\description{
Aggregate cells in bulk data based on the Gaussian kernal
}
