% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTrainingSet_functions.R
\name{BulkCells_subfxn}
\alias{BulkCells_subfxn}
\title{Sample cells}
\usage{
BulkCells_subfxn(i, cell.prob, max.cent, tot.cells, n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Sample cells
}
