% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_gps}
\alias{cell_gps}
\title{Cell type cluster IDs}
\format{
A character vector with 500 cluster IDs
}
\usage{
cell_gps
}
\description{
Cell type cluster IDs for the counts_gps single-cell data.
}
\examples{
data(cell_gps)
}
\keyword{datasets}
