% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTrainingSet_functions.R
\name{CellProb}
\alias{CellProb}
\title{Calculate cell probability}
\usage{
CellProb(parameters, latent_distance, max.cent, tot.cells)
}
\arguments{
\item{tot.cells}{}
}
\value{

}
\description{
Calculate the probability of each cell based on a Gaussian distribution
}
