% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeProximityScore.R
\name{ComputeProximityScore}
\alias{ComputeProximityScore}
\title{ComputeProximityScore}
\usage{
ComputeProximityScore(cond)
}
\arguments{
\item{cond}{ConDecon object}
}
\value{
ConDecon object with computed proximity score
}
\description{
Compute the average distance between the query dataset and the nearest 10 training datasets normalized by the average distance between training datasets
}
