% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2Latent_functions.R
\name{CorBulk}
\alias{CorBulk}
\title{CorBulk}
\usage{
CorBulk(bulk, genes, gene.index)
}
\arguments{
\item{bulk}{Matrix of query bulk data (features x samples)}

\item{genes}{Character vector of variable features from single-cell data}

\item{gene.index}{Numeric vector of gene rank}
}
\value{
Matrix of rank correlation for each bulk sample
}
\description{
CorBulk
}
