% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts_gps}
\alias{counts_gps}
\title{Example single-cell counts data}
\format{
A matrix with 5,000 rows (genes) and 500 columns (cells):
}
\source{
{Simulated single-cell data generated by Splatter} \url{https://doi.org/10.1186/s13059-017-1305-0}
}
\usage{
counts_gps
}
\description{
Example single-cell gene expression counts data containing 9 discrete cell types/groups (gps)
generated by Splatter.
}
\examples{
data(counts_gps)
}
\keyword{datasets}
