% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTrainingSet_functions.R
\name{Find_Sigma}
\alias{Find_Sigma}
\title{Find sigma}
\usage{
Find_Sigma(sigma, x_i, width)
}
\arguments{
\item{sigma}{variable}

\item{x_i}{range of sigma values}

\item{width}{standard deviation of Gaussian}
}
\value{
Sigma value
}
\description{
Find sigma
}
