% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2Latent_functions.R
\name{GeneIndex}
\alias{GeneIndex}
\title{GeneIndex}
\usage{
GeneIndex(knn, count, variable.features, k)
}
\arguments{
\item{knn}{Matrix of cells with their nearest neighbor}

\item{count}{Matrix of single-cell count data}

\item{variable.features}{Character vector of variable features from single-cell data}

\item{k}{Number of nearest neighbor cells aggregated together when calculating rank correlation}
}
\value{
Matrix of rank indices for variable features ordered by expression
}
\description{
GeneIndex
}
