% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2Latent_functions.R
\name{GeneIndex_subfxn}
\alias{GeneIndex_subfxn}
\title{GeneIndex_subfxn}
\usage{
GeneIndex_subfxn(i, geneID)
}
\arguments{
\item{i}{Bulk sample}

\item{geneID}{Vector with index of variable features}
}
\value{
Numeric vector of gene rank
}
\description{
GeneIndex_subfxn
}
