% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{latent_gps}
\alias{latent_gps}
\title{Example PCA latent space}
\format{
A matrix with 500 rows (cells) and 50 columns (principal components)
}
\usage{
latent_gps
}
\description{
Example PCA latent space associated with the counts_gps single-cell data.
}
\examples{
data(latent_gps)
}
\keyword{datasets}
