% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2Latent_functions.R
\name{Map2Latent}
\alias{Map2Latent}
\title{Map2Latent}
\usage{
Map2Latent(TrainingSet, latent, count, bulk, variable.features, k = 1)
}
\arguments{
\item{TrainingSet}{Training set data (output from BuildTrainingSet)}

\item{latent}{matrix of single-cell latent space (cells x dims)}

\item{count}{single-cell count matrix (features x cells)}

\item{bulk}{matrix of query bulk data (features x samples)}

\item{variable.features}{character vector of variable features}

\item{k}{Number of nearest neighbor cells aggregated together when calculating rank correlation}
}
\value{
ConDecon object with low dimensional embedding of the space
of cell abundances and correlations
}
\description{
Use the latent space to expand the distributions
of cell abundances and correlations on a basis of functions
}
\examples{
data(counts_gps)
data(latent_gps)
data(bulk_gps)
data(variable_genes_gps)

# For this example, we will reduce the training size to max.iter = 50 to reduce run time
TrainingSet = BuildTrainingSet(count = counts_gps, latent = latent_gps, max.iter = 50)

ConDecon_obj = Map2Latent(TrainingSet = TrainingSet, latent = latent_gps, count = counts_gps,
bulk = bulk_gps, variable.features = variable_genes_gps)
}
