% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta_data_gps}
\alias{meta_data_gps}
\title{meta data for 'counts_gps', the single-cell RNA-seq data}
\format{
A data frame with 3 columns
}
\usage{
meta_data_gps
}
\description{
data frame that contains the UMAP coordinates of each cell and cluster IDs
}
\examples{
data(meta_data_gps)
}
\keyword{datasets}
