% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildTrainingSet_functions.R
\name{PickCells}
\alias{PickCells}
\title{Pick cells based on the Gaussian kernal}
\usage{
PickCells(max.iter, step, cell.prob, max.cent, tot.cells, n, verbose)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Pick cells based on the Gaussian kernal
}
