% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConDecon.R
\name{PlotConDecon}
\alias{PlotConDecon}
\title{Plot ConDecon_obj}
\usage{
PlotConDecon(
  ConDecon_obj,
  umap = ConDecon_obj$TrainingSet$latent[, 1:2],
  samples = NULL,
  pt.size = 1,
  cells = NULL,
  title_names = samples
)
}
\arguments{
\item{ConDecon_obj}{ConDecon object (output from RunConDecon)}

\item{umap}{2-dimensional coordinates of reference single-cell data (default = first 2 dimensions of latent space)}

\item{samples}{Vector of query bulk sample(s) to plot (default = NULL, plot all samples)}

\item{pt.size}{size of the points plotted (default = 1)}

\item{cells}{Vector of cells from the reference single-cell data to plot (default = NULL, plot all cells)}

\item{title_names}{Title of each plot (default = samples, use NULL for no title)}
}
\value{
Plot the z-score cell probability values for each query bulk sample
}
\description{
Plot ConDecon's predicted relative cell probability z-scores for each sample
}
\examples{
data(counts_gps)
data(latent_gps)
data(bulk_gps)
data(variable_genes_gps)

# For this example, we will reduce the training size to max.iter = 50 to reduce run time
ConDecon_obj = RunConDecon(counts = counts_gps, latent = latent_gps, bulk = bulk_gps,
variable.features = variable_genes_gps, max.iter = 50)

# Plot ConDecon relative cell prob
PlotConDecon(ConDecon_obj)
}
