% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictCellProb.R
\name{PredictCellProb}
\alias{PredictCellProb}
\title{PredictCellProb}
\usage{
PredictCellProb(bulk, count, variable.features, output)
}
\arguments{
\item{bulk}{matrix of the query bulk data (features x samples)}

\item{count}{single-cell count matrix (features x cells)}

\item{variable.features}{character vector of the most variable features}

\item{output}{ConDecon object with fit model}
}
\value{
ConDecon object with inferred cell probabilities for each query bulk sample
}
\description{
Using the fit model, predict the cell abundance distributions
}
\examples{
data(counts_gps)
data(variable_genes_gps)
data(bulk_gps)

TrainingSet = BuildTrainingSet(count = counts_gps, latent = latent_gps)
ConDecon_obj = Map2Latent(output = TrainingSet, latent = latent_gps, count = counts_gps,
bulk = bulk_gps, variable.features = variable_genes_gps)
ConDecon_obj = BuildModel(ConDecon_obj)

ConDecon_obj = PredictCellProb(bulk = bulk_gps, count = counts_gps,
variable.features = variable_genes_gps, output = ConDecon_obj)
}
