% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunConDecon.R
\name{RunConDecon}
\alias{RunConDecon}
\title{Run ConDecon}
\usage{
RunConDecon(
  counts,
  latent,
  bulk,
  variable.features,
  max.iter = 5000,
  max.cent = 5,
  dims = 10,
  degree = 1,
  step = ifelse(max.iter <= 10000, max.iter, 10000),
  min.cent = 1,
  n = round(ncol(counts)/2),
  trainingset = NULL,
  sigma_min_cells = NULL,
  sigma_max_cells = NULL,
  verbose = FALSE
)
}
\arguments{
\item{counts}{single-cell counts matrix (features x cells)}

\item{latent}{matrix of single-cell latent space (cells x dims)}

\item{bulk}{matrix of query bulk data (features x samples)}

\item{variable.features}{character vector of the most variable features}

\item{max.iter}{size of the training dataset (default = 5,000)}

\item{max.cent}{max number of centers in the Gaussian (default = 5)}

\item{dims}{number of dimensions from latent (default = 10)}

\item{degree}{degree of the polynomial used}

\item{step}{manually parallelize building the training dataset}

\item{min.cent}{min number of centers in the Gaussian (default = 1)}

\item{n}{number of cells to be chosen to create the training dataset (default
is half the number of cells in the counts matrix)}

\item{trainingset}{pre-generated training dataset}

\item{sigma_min_cells}{min number of cells that should be captured by the standard
deviation of the Gaussian}

\item{sigma_max_cells}{max number of cells that should be captured by the standard
deviation of the Gaussian}

\item{verbose}{logical indicating whether to print progress (default = FALSE)}
}
\value{
ConDecon object with continuous deconvolution results
}
\description{
Run ConDecon (continuous deconvolution) to estimate the cell abundances associated with each query bulk sample
}
\examples{
data(counts_gps)
data(latent_gps)
data(bulk_gps)
data(variable_genes_gps)

ConDecon_obj = RunConDecon(counts = counts_gps, latent = latent_gps,
bulk = bulk_gps, variable.features = variable_genes_gps)
}
