% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFeatures.R
\name{set_zero}
\alias{set_zero}
\title{Set values to zero}
\usage{
set_zero(v, threshold)
}
\arguments{
\item{v}{Vector of predicted cell probability values for a single sample}

\item{threshold}{Numeric of the distribution that represents the determined quartile value}
}
\value{
Vector of cell probability values truncated based on threshold
}
\description{
Set values to zero
}
