% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransferFeatures.R
\name{smooth_cell_prob}
\alias{smooth_cell_prob}
\title{Smooth ConDecon's cell.prob matrix}
\usage{
smooth_cell_prob(cell_prob, probs = 0.75)
}
\arguments{
\item{cell_prob}{Matrix of ConDecon's predicted cell probability distributions}

\item{probs}{Numeric value indicating the bottom quartile of the distribution that should be removed for the purpose of smoothing long tails of the predicted cell probability distribuiton (default = 0.75)}
}
\value{
Matrix with truncated cell probability values based on quartile value
}
\description{
Smooth ConDecon's cell.prob matrix
}
