% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_prop_gps}
\alias{true_prop_gps}
\title{True cell type proportions}
\format{
A matrix with 5 rows (5 bulk samples) and 9 columns (number of cell types)
}
\usage{
true_prop_gps
}
\description{
True cell type proportions associated with the bulk_gps bulk data.
}
\examples{
data(true_prop_gps)
}
\keyword{datasets}
