% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{umap_embedding_gps}
\alias{umap_embedding_gps}
\title{Example UMAP 2D embedding}
\format{
A matrix with 500 rows (cells) and 2 columns (UMAP)
}
\usage{
umap_embedding_gps
}
\description{
Example UMAP 2D embedding of the latent space (latent_gps).
}
\examples{
data(umap_embedding_gps)
}
\keyword{datasets}
