% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_genes_gps}
\alias{variable_genes_gps}
\title{Variable genes}
\format{
A character vector containing 2,000 variable genes
}
\usage{
variable_genes_gps
}
\description{
2,000 variable genes associated with the counts_gps single-cell data.
}
\examples{
data(variable_genes_gps)
}
\keyword{datasets}
