% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotConDecon.R
\name{PlotConDecon}
\alias{PlotConDecon}
\title{Plot ConDecon_obj}
\usage{
PlotConDecon(
  ConDecon_obj,
  umap = ConDecon_obj$TrainingSet$latent[, 1:2],
  samples = NULL,
  pt.size = 1,
  rm_outliers = FALSE,
  cells = NULL,
  title_names = NULL
)
}
\arguments{
\item{ConDecon_obj}{ConDecon object (output from RunConDecon)}

\item{umap}{2-dimensional coordinates of reference single-cell data (default = first 2 dimensions of latent space)}

\item{samples}{character vector of the name(s) of the query bulk sample(s) to plot; relative probability will still be calculated between all sample(s) (default = NULL)}

\item{pt.size}{size of the points plotted (default = 1)}

\item{rm_outliers}{Boolean indicating whether to remove outlier cell probability values from the z-score calculation based on +- 1.5*IQR (default = FALSE)}

\item{cells}{vector of cell names/index indicating which cells from the reference single-cell data to plot (default = NULL, plot all cells)}

\item{title_names}{title of the plot (default = NULL, use bulk sample names from ConDecon object)}
}
\value{
Plot the z-score cell probability values for each query bulk sample
}
\description{
Calculate and plot the cell probability z-score for each cell and bulk sample. Cell probability z-score associated with query sample_i:
z-score_i = [ConDecon_obj$norm_cell.prob[,i] - rowMeans(ConDecon_obj$norm_cell.prob)] / RowSds(ConDecon_obj$norm_cell.prob)
}
\examples{
data(counts_gps)
data(latent_gps)
data(bulk_gps)
data(variable_genes_gps)

ConDecon_obj = RunConDecon(counts = counts_gps, latent = latent_gps, bulk = bulk_gps,
variable.features = variable_genes_gps)

# add z-score cell abundances to ConDecon object: ConDecon$relative_cell.prob
ConDecon_obj = PlotConDecon(ConDecon_obj)

# Plot the z-score cell abundances for the query bulk sample(s)
PlotConDecon(ConDecon_obj)
}
